Remove unnecessary to_path_buf() in pattern
